fix: argument type mismatch 해결

+++ src/main/java/com/server/ggini/domain/restaurant/dto/response/SeoulConfirmRestaurantGetResponse.java
@@ -4,6 +4,6 @@ public record SeoulConfirmRestaurantGetResponse(
         Long id,
         String name,
         String roadAddress,
-        Long distance
+		Double distance // Long -> argument type mismatch
 ) {
 }
